[PDB] Read inc-link thunks from /* LINKER */ module and create symbols
Summary:
The "/* LINKER */" module contains symbols for thunks used for incremental
linking. A `callq` to a symbol gets statically linked to a thunk that is just a
jmp to the actual target symbol. This information is contained in this "/*
LINKER */" module.

In order for ThreadPlans to handle stepping through these thunks there has to be
something pertaining to it returned from ResolveSymbolContext.  So while looking
up symbols in the PDB by VA we can create a Symbol from the PDB Symbol type
which was previously ignored and then the ThreadPlan can properly handle this
step.

Subscribers: jdoerfert

Differential Revision: https://reviews.llvm.org/D63526