[NativePDB] Change GetOrCreateBlockDecl to private and introduce CreateBlockDecl
Summary:
We're removing all clang specific details from PdbAstBuilder in order to
make it an abstract base class and then create subclasses specific to
the language.

GetOrCreateBlockDecl returns a clang::BlockDecl. But the only class that
uses the return is PdbAstBuilder itself. So we can simply create a
wrapper with no return that jumps to the actual creation block in order
to remove the clang specific behavior from PdbAstBuilder.

Differential Revision: https://reviews.llvm.org/D64693